requirements.txt
einops==0.7.0
ftfy==6.1.3
numpy==1.26.4
Pillow==10.2.0
regex==2023.12.25
scipy==1.12.0
setuptools==69.1.1
tensorboardX==2.6.2.2
timm==0.9.16
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
scipy 
scikit-learn
streamlit
gradio
matplotlib
seaborn
huggingface_hub
onnx==1.17.0
onnxruntime==1.21.0